Wed, Jul 20, 2022, 12:48 PM +00:00

+++ average.c
@@ -1,23 +1,23 @@
 #include <stdio.h>
 
-float average (float array[]);
+double average (double array[]);
 
 int main(void)
 {
     // Do some averages
-    float array1[] = {2, 1, 2};
-    float array2[] = {7, 6, 6};
-    float array3[] = {8, 5, 7};
+    double array1[] = {2, 1, 2};
+    double array2[] = {7, 6, 6};
+    double array3[] = {8, 5, 7};
 
-    printf("Array1: %3f\n", average(array1));
-    printf("Array2: %f\n", average(array2));
-    printf("Array3: %f\n", average(array3));
+    printf("Array1: %d\n", average(array1));
+    printf("Array2: %d\n", average(array2));
+    printf("Array3: %d\n", average(array3));
 }
 
-float average (float array[])
+double average (double array[])
 {
-    float len = sizeof(*array) / sizeof(int);
-    float sum = 0;
+    double len = sizeof(*array) / sizeof(int);
+    double sum = 0;
 
     for (int i = 0; i < len; i ++)
     {
